Refactor routes code

+++ 1_Requests_Starter/backend/flaskr/__init__.py
@@ -9,6 +9,17 @@ from models import setup_db, Book
 
 BOOKS_PER_SHELF = 8
 
+
+def paginate_books(selection):
+    page = request.args.get("page", 1, type=int)
+    start = (page - 1) * BOOKS_PER_SHELF
+    end = start + BOOKS_PER_SHELF
+    formatted_books = [book.format() for book in selection]
+    current_books = formatted_books[start:end]
+
+    return current_books
+
+
 # @TODO: General Instructions
 #   - As you're creating endpoints, define them and then search for 'TODO' within the frontend to update the endpoints there.
 #     If you do not update the endpoints, the lab will not work - of no fault of your API code!
@@ -40,18 +51,18 @@ def create_app(test_config=None):
     #         Response body keys: 'success', 'books' and 'total_books'
     # TEST: When completed, the webpage will display books including title, author, and rating shown as stars
     @app.route("/books")
-    def get_books():
-        page = request.args.get("page", 1, type=int)
-        start = (page - 1) * BOOKS_PER_SHELF
-        end = start + BOOKS_PER_SHELF
-        books = Book.query.order_by("id").all()
-        formatted_books = [book.format() for book in books]
+    def retrieve_books():
+        selection = Book.query.order_by("id").all()
+        current_books = paginate_books(selection)
+
+        if len(current_books) == 0:
+            abort(404)
 
         return jsonify(
             {
                 "success": True,
-                "books": formatted_books[start:end],
-                "total_books": len(formatted_books),
+                "books": current_books,
+                "total_books": len(selection),
             }
         )
 
@@ -60,80 +71,83 @@ def create_app(test_config=None):
     #         and should follow API design principles regarding method and route.
     #         Response body keys: 'success'
     # TEST: When completed, you will be able to click on stars to update a book's rating and it will persist after refresh
-    @app.route("/books/<int:book_id>/update-rating", methods=["PATCH"])
-    def update_book_rating(book_id):
-
-        rating = request.get_json()["rating"]
-        book = Book.query.get(book_id)
-        error_inserting_db = False
+    @app.route("/books/<int:book_id>", methods=["PATCH"])
+    def update_book(book_id):
+        body = request.get_json()
 
         try:
-            book.rating = rating
-            book.update()
-        except:
-            book.rollback()
-            error_inserting_db = True
+            book = Book.query.filter(Book.id == book_id).one_or_none()
 
-        if error_inserting_db:
-            abort(500)
-        else:
-            return jsonify({"success": True})
+            if book is None:
+                abort(404)
+
+            if "rating" in body:
+                book.rating = int(body.get("rating"))
+
+            book.update()
+
+            return jsonify({"success": True, "id": book.id})
+
+        except:
+            abort(400)
 
     # @TODO: Write a route that will delete a single book.
     #        Response body keys: 'success', 'deleted'(id of deleted book), 'books' and 'total_books'
     #        Response body keys: 'success', 'books' and 'total_books'
-
     # TEST: When completed, you will be able to delete a single book by clicking on the trashcan.
     @app.route("/books/<int:book_id>/delete", methods=["DELETE"])
     def delete_book(book_id):
-        book = Book.query.get(book_id)
-        error = False
-
         try:
+            book = Book.query.filter(Book.id == book_id).one_or_none()
+
+            if book is None:
+                abort(404)
+
             book.delete()
+            selection = Book.query.order_by("id").all()
+            current_books = paginate_books(selection)
+
+            return jsonify(
+                {
+                    "success": True,
+                    "deleted": book_id,
+                    "books": current_books,
+                    "total_books": len(selection),
+                }
+            )
 
         except:
-            book.rollback()
-            error = True
-
-        if error:
-            abort(500)
-        else:
-            return jsonify({"success": True})
+            abort(422)
 
     # @TODO: Write a route that create a new book.
     #        Response body keys: 'success', 'created'(id of created book), 'books' and 'total_books'
     # TEST: When completed, you will be able to a new book using the form. Try doing so from the last page of books.
     #       Your new book should show up immediately after you submit it at the end of the page.
-    @app.route("/books/create", methods=["POST"])
+    @app.route("/books", methods=["POST"])
     def create_book():
-        error = False
+        body = request.get_json()
+
+        new_title = body.get("title", None)
+        new_author = body.get("author", None)
+        new_rating = body.get("rating", None)
 
         try:
-            title = request.get_json()["title"]
-            author = request.get_json()["author"]
-            rating = request.get_json()["rating"]
-
-            book = Book(title=title, author=author, rating=rating)
+            book = Book(title=new_title, author=new_author, rating=int(new_rating))
             book.insert()
 
-        except:
-            book.rollback()
-            error = True
-
-        if error:
-            abort(500)
-        else:
-            books = Book.query.order_by("id").all()
-            formatted_books = [book.format() for book in books]
+            selection = Book.query.order_by("id").all()
+            current_books = paginate_books(selection)
 
             return jsonify(
                 {
                     "success": True,
                     "created": book.id,
-                    "books": formatted_books,
-                    "total_books": len(books),
+                    "books": current_books,
+                    "total_books": len(selection),
                 }
             )
+
+        except:
+            abort(422)
 
     return app
